requirements.txt
Pillow~=8.4.0
ipython~=7.22.0
pip~=21.2.4
toml~=0.10.2
tornado~=6.1
Django~=4.0.2
sqlparse~=0.4.2
asgiref~=3.5.0
colorama~=0.4.4
Jinja2~=3.0.3
numpy~=1.20.0
setuptools~=49.2.1
